requirements.txt
fastapi==0.115.12
numpy==2.0.2
pillow==11.2.1
pyngrok==7.2.5
python-multipart==0.0.20
tensorflow==2.18.0
tensorflow-datasets==4.9.8
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.1
tensorflow-probability==0.25.0
tensorflow-text==2.18.1
tensorflow_decision_forests==1.11.0
uvicorn==0.34.2
